cmd/server: move test (live) searcher to '*_test.go' files
